backend/requirements.txt
# Core web framework
fastapi==0.110.1
starlette==0.37.2
uvicorn==0.25.0
python-multipart==0.0.20
python-dotenv==1.1.1

# Database
motor==3.3.1
pymongo==4.5.0

# Authentication & security
bcrypt==4.1.3
PyJWT==2.10.1
cryptography==46.0.3
passlib==1.7.4
slowapi==0.1.9

# Data validation
pydantic==2.12.3
pydantic_core==2.41.4
annotated-types==0.7.0
email-validator==2.3.0

# HTTP clients
httpx==0.27.0
requests==2.32.5
aiohttp==3.13.2
aiofiles==25.1.0

# Image processing
pillow==12.0.0
pillow_heif==1.1.1

# Cloud storage
boto3==1.40.55
botocore==1.40.55
s3transfer==0.14.0

# Monitoring & error tracking
psutil==7.1.2
sentry-sdk==2.48.0

# Antivirus scanning (optional — ClamAV must be installed)
pyClamd==0.4.0

# CLI tools
typer==0.20.0
rich==14.2.0
click==8.3.0

# Transitive dependencies (pinned for reproducibility)
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
Deprecated==1.2.18
dnspython==2.8.0
ecdsa==0.19.1
frozenlist==1.8.0
h11==0.16.0
idna==3.11
jmespath==1.0.1
limits==5.6.0
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
oauthlib==3.3.1
packaging==25.0
propcache==0.4.1
pyasn1==0.6.1
pycparser==2.23
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests-oauthlib==2.0.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchfiles==1.1.1
webencodings==0.5.1
wrapt==1.17.3
yarl==1.22.0

# ──────────────────────────────────────────
# Removed from production requirements:
# - black, flake8, isort, mypy, pyflakes, pycodestyle, ruff, pytest
#   (dev-only tools — use requirements-dev.txt)
# - python-jose (unmaintained, CVEs — PyJWT is already used)
# - numpy, pandas (unused — remove ~100MB from Docker image)
# - bleach (deprecated Jan 2023)
# - jq, s5cmd, pathspec, pluggy, iniconfig, pytokens (unused)
